requirements.txt
aiogram==2.23.1
python-dotenv
sqlalchemy